test(pdf): move the credentialed probe into a *.live.test.ts file
ClawSweeper P1: ordinary test discovery collected an
evidence-credential-bearing probe. The repository convention excludes
*.live.test.ts from every project, so the configured-provider case now
lives there and runs only when explicitly invoked with credentials.
